AutoCollapse plugin
added AutoCollapse plugin. add the attribute of autoCollapse to an element and if that element has no child elements it will be removed from the DOM this also adds a class of "yes-[element id]" when the element is in the dom and "no-[element id]" when the element is removed from the dom. This will make selective CSS for this work with dom changes as elements get removed from the dom.